refactor code used by generate_pipeline
